chore: refactor code to reflect the changes in namespace
