+++ app/src/main/java/com/github/only52607/compose/window/app/ui/theme/Theme.kt
@@ -1,4 +1,4 @@
-package com.github.only52607.compose.window.ui.theme
+package com.github.only52607.compose.window.app.ui.theme
 
 import android.app.Activity
 import android.os.Build
