+++ app/src/main/java/com/github/only52607/compose/window/app/ui/theme/Type.kt
@@ -1,4 +1,4 @@
-package com.github.only52607.compose.window.ui.theme
+package com.github.only52607.compose.window.app.ui.theme
 
 import androidx.compose.material3.Typography
 import androidx.compose.ui.text.TextStyle
